+++ src/BrpProxy/Mappers/GeslachtHelpers.cs
@@ -1,0 +1,19 @@
+﻿using BrpProxy.Generated;
+using HaalCentraal.BrpProxy.Generated;
+
+namespace BrpProxy.Mappers;
+
+public static class GeslachtHelpers
+{
+    public static bool IsVrouw(this IWaardetabel? geslacht) => geslacht?.Code.ToUpperInvariant() == "V";
+
+    public static bool IsVrouw(this NaamPersoon naam) => naam.Geslacht.IsVrouw();
+
+    public static bool IsMan(this IWaardetabel? geslacht) => geslacht?.Code.ToUpperInvariant() == "M";
+
+    public static bool IsMan(this AbstractPartner? partner) =>
+        partner is Partner p &&
+        p.Geslacht.IsMan();
+
+    public static bool IsMan(this NaamPersoon naam) => naam.Geslacht.IsMan();
+}
