+++ src/BrpProxy/Mappers/VolledigeNaamMapper.cs
@@ -1,0 +1,16 @@
+﻿using BrpProxy.Generated;
+using HaalCentraal.BrpProxy.Generated;
+using System.Text.RegularExpressions;
+
+namespace BrpProxy.Mappers;
+
+public static class VolledigeNaamMapper
+{
+    public static string VolledigeNaam(this NaamPersoon naam, IWaardetabel? geslacht)
+    {
+        var adellijkeTitel = naam.AdellijkeTitelPredicaat.MapNaarAdellijkeTitel(geslacht);
+        var predikaat = naam.AdellijkeTitelPredicaat.MapNaarPredicaat(geslacht);
+
+        return Regex.Replace($"{predikaat} {naam.Voornamen} {adellijkeTitel} {naam.Voorvoegsel} {naam.Geslachtsnaam}", @"\s+", " ").Trim();
+    }
+}
